changed doub prec to single prec

+++ gpumcml_opencl_simplified/mem.cpp
@@ -123,13 +123,13 @@ int InitSimStates(SimState* HostMem, SimulationStruct* sim, cl_context context, 
   }
 
   // output arrays absorption and transmittance
-  size = (sim->nTetras+1) * sizeof(UINT64);
+  size = (sim->nTetras+1) * sizeof(UINT32);
   *absorption_mem_obj = clCreateBuffer(context, CL_MEM_READ_WRITE, size, NULL, &ret);
   if(ret!=CL_SUCCESS){
     printf("Error creating absorption mem buffer, exiting\n");
     exit(-1);
   }
-  HostMem->absorption = (UINT64*)malloc(size);
+  HostMem->absorption = (UINT32*)malloc(size);
   for(int i = 1; i < sim->nTetras+1; i++)
   {
     HostMem->absorption[i] = 0;
@@ -140,13 +140,13 @@ int InitSimStates(SimState* HostMem, SimulationStruct* sim, cl_context context, 
     exit(-1);
   }
 
-  size = (sim->nTetras)* 4 * sizeof(UINT64);
+  size = (sim->nTetras)* 4 * sizeof(UINT32);
   *transmittance_mem_obj = clCreateBuffer(context, CL_MEM_READ_WRITE, size, NULL, &ret);
   if(ret!=CL_SUCCESS){
     printf("Error creating transmittance mem buffer, exiting\n");
     exit(-1);
   }
-  HostMem->transmittance = (UINT64*)malloc(size);
+  HostMem->transmittance = (UINT32*)malloc(size);
   for(int i = 0; i < (sim->nTetras)* 4; i++)
   {
     HostMem->transmittance[i] = 0;
@@ -252,13 +252,13 @@ int InitSimStates(SimState* HostMem, SimulationStruct* sim, cl_context context, 
 int CopyDeviceToHostMem(SimState* HostMem, SimulationStruct* sim, cl_command_queue command_queue, cl_mem x_mem_obj, cl_mem absorption_mem_obj, cl_mem transmittance_mem_obj, cl_mem debug_mem_obj)
 {
   cl_int ret;
-  ret = clEnqueueReadBuffer(command_queue, absorption_mem_obj, CL_TRUE, 0, (sim->nTetras+1)*sizeof(UINT64), HostMem->absorption, 0, NULL, NULL);
+  ret = clEnqueueReadBuffer(command_queue, absorption_mem_obj, CL_TRUE, 0, (sim->nTetras+1)*sizeof(UINT32), HostMem->absorption, 0, NULL, NULL);
   if(ret != CL_SUCCESS){
     printf("Error reading absorption buffer, exiting\n");
     exit(-1);
   }
 
-  ret = clEnqueueReadBuffer(command_queue, transmittance_mem_obj, CL_TRUE, 0, (sim->nTetras)*4*sizeof(UINT64), HostMem->transmittance, 0, NULL, NULL);
+  ret = clEnqueueReadBuffer(command_queue, transmittance_mem_obj, CL_TRUE, 0, (sim->nTetras)*4*sizeof(UINT32), HostMem->transmittance, 0, NULL, NULL);
   if(ret != CL_SUCCESS){
     printf("Error reading transmittance buffer, exiting\n");
     exit(-1);
